# openapi-action.yaml
openapi: 3.1.0
info:
  title: NeginAI Secure SQL Gateway
  version: 1.3.0
  description: Secure NeginPakhsh business assistant for live reports, recurring automations, conditional alerts, and Admin planning scenarios.
servers:
  - url: https://ai.neginpakhsh.com
    description: Stable HTTPS endpoint for NeginAI.
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    neginOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://ai.neginpakhsh.com/oauth/authorize
          tokenUrl: https://ai.neginpakhsh.com/oauth/token
          scopes:
            company.read: Read authorized NeginPakhsh company reports
  schemas:
    ChatRequest:
      type: object
      required: [message]
      additionalProperties: false
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 4000
          description: >-
            The user's complete Persian request. This includes internal reports and every
            automation request such as every day, every hour, at a specific time, remind me,
            run automatically, or notify me when a condition becomes true. Send the original
            request without rewriting or dropping its schedule or condition.
            Admin requests about budgets, forecasts, planning scenarios, assumptions,
            targets, approvals, locks, or variance analysis must also be sent unchanged.
        conversation_id:
          type: [string, 'null']
          maxLength: 100
          description: >-
            Stable id for this ChatGPT thread. On the first call generate one UUID and send it.
            On every later call in the same ChatGPT conversation, reuse the exact
            conversation_id returned by the previous call. Never generate a new id for a follow-up.
    ChatResponse:
      type: object
      required: [conversation_id, answer]
      properties:
        conversation_id: {type: string}
        answer:
          type: string
          description: Final, user-facing Persian answer produced from authorized company data.
        columns:
          type: array
          items: {type: string}
        rows:
          type: array
          items:
            type: array
            items: {}
        row_count: {type: integer}
        execution_time: {type: [number, 'null']}
        truncated: {type: boolean}
        sources:
          type: array
          items: {type: string}
        sql: {type: [string, 'null']}
        clarification_required: {type: boolean}
...
security:
  - neginOAuth: [company.read]
  - apiKeyAuth: []
paths:
  /chat:
    post:
      operationId: chatWithNeginAI
      x-openai-isConsequential: false
      summary: Answer company questions and manage automations plus Admin planning scenarios
      description: >-
        Mandatory primary action for every request about company sales, branches,
        customers, products, inventory, finance, treasury, reports, definitions,
        budgets, forecasts, planning scenarios, targets, assumptions, approvals and variance analysis,
        database information, recurring work, scheduled reports, reminders, or
        condition-based alerts. If the user says every day, every hour, periodically,
        automatically, at a specific time, remind me, or notify me when something
        happens, call this action immediately. Never claim that scheduling is unavailable
        and never tell the user to return and ask again at the scheduled time: NeginAI's
        backend owns scheduling and execution. Send the user's complete original request
        before answering. Each user should authenticate with NeginAI OAuth. Never request
        a password in chat and never replace this action with web search. Keep one stable
        conversation_id across all calls in the same ChatGPT conversation so follow-up
        filters, dates, and automation references are inherited correctly. Phrases such as
        "همین فروش", "همین گزارش", or "همان نتیجه" must reuse that id and the backend answer
        must be returned without recalculating or inventing figures in the GPT layer.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChatRequest'}
      responses:
        '200':
          description: Final answer and optional structured report data
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ChatResponse'}
        '401': {description: Action authentication failed}
        '502': {description: Internal report generation failed}
  /entities/search:
    get:
      operationId: resolveBusinessEntity
      x-openai-isConsequential: false
      summary: Resolve a brand or manufacturer name; always use this before composing sales SQL
      description: Distinguishes brands from manufacturers and returns canonical IDs plus correct join guidance.
      parameters:
        - {name: q, in: query, required: true, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 10}}
      responses:
        '200': {description: Matching brands/manufacturers and technical join guidance}
...